backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.9

# Database and storage
qdrant-client==1.10.1

# Data validation and serialization
pydantic==2.9.2
pydantic-settings==2.4.0

# CORS and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# WebSocket support
websockets==12.0

# Environment variables
python-dotenv==1.0.1

# Development tools
httpx==0.27.0
pytest==8.3.2
pytest-asyncio==0.24.0

# Utilities
numpy==1.26.4